# resources/swagger.yaml
---
swagger: "2.0"
info:
  description: "Issue Manager API"
  version: "1.0.0"
  title: "Issue Manager API"
host: "localhost:3000"
basePath: "/"
schemes:
- "http"
- "https"
produces:
- "application/json"
paths:
  /issue:
    x-controller: "issueController"
    get:
      tags:
        - "Issue"
      summary: "get issue"
      description: "Get Status"
      operationId: "getIssue"
      parameters:
        - name: id
          in: query
          type: string
          required: false
      responses:
        200:
          description: array of issues
          schema:
            type: array
            items:
              $ref: "#/definitions/Issue"
        404:
          $ref: "#/responses/NotFound"  
        500:
          $ref: "#/responses/UnexpectedError"
    post:
      tags:
        - "Issue"
      summary: "create issue"
      description: "create issue"
      operationId: "createIssue"
      parameters:
        - name: issue
          in: body
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
              userId:
                type: string                
      
      responses:
        200:
          description: "Successful"
          schema:
            $ref: "#/definitions/Issue"
        500:
          $ref: "#/responses/UnexpectedError"
  /issue/list:
    x-controller: "issueController"
    get:
      tags:
        - "Issue"
      summary: "get issue"
      description: "Get Status"
      operationId: "getIssueList"
      parameters:
        - name: agentId
          in: query
          type: string
          required: false
        - name: state
          type: string
          enum: [UNASSIGNED, ASSIGNED, RESOLVED]
          in: query
        - name: userId
          in: query
          type: string
          required: false
      responses:
        200:
          description: array of issues
          schema:
            type: array
            items:
              $ref: "#/definitions/Issue"
        500:
          $ref: "#/responses/UnexpectedError"
          
  /agent:
    x-controller: "agentController"
    post:
      tags:
        - "Agent"
      summary: "creates a new support agent"
      description: "creates a new support agent"
      operationId: "createAgent"
      parameters:
        - name: agent
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
      responses:
        200:
          description: "Successful"
          schema:
            $ref: "#/definitions/Agent"
        500:
          $ref: "#/responses/UnexpectedError"
    get:
      tags:
        - "Agent"
      summary: "get Agent"
      description: "Get Agent"
      operationId: "getAgent"
      parameters:
        - name: id
          in: query
          type: string
          required: true
      responses:
        200:
          description: Agent
          schema:
            $ref: "#/definitions/Agent"
        404:
          $ref: "#/responses/NotFound"  
        500:
          $ref: "#/responses/UnexpectedError"
  /agent/list:
    x-controller: "agentController"
    get:
      tags:
        - "Agent"
      summary: "get Agent List"
      description: "Get Agent List"
      operationId: "getAgentList"
      parameters:
        - name: state
          type: string
          enum: [FREE, BUSY]
          in: query
      responses:
        200:
          description: array of agents
          schema:
            type: array
            items:
              $ref: "#/definitions/Agent"
        500:
          $ref: "#/responses/UnexpectedError"

  /user:
    x-controller: "userController"
    get:
      tags:
        - "User"
      summary: "get User"
      description: "Get User"
      operationId: "getUser"
      parameters:
        - name: id
          in: query
          type: string
          required: true
      responses:
        200:
          description: Agent
          schema:
            $ref: "#/definitions/Agent"
        404:
          $ref: "#/responses/NotFound"  
        500:
          $ref: "#/responses/UnexpectedError"    
    post:
      tags:
        - "User"
      summary: "Create a new user"
      description: "Create a new user"
      operationId: "createUser"
      parameters:
        - name: user
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
      responses:
        200:
          description: "Successful"
          schema:
            $ref: "#/definitions/User"          
        500:
          $ref: "#/responses/UnexpectedError"
  /user/list:
    x-controller: "userController"
    get:
      tags:
        - "User"
      summary: "get User List"
      description: "Get User List"
      operationId: "getUserList"
      responses:
        200:
          description: Agent
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
        500:
          $ref: "#/responses/UnexpectedError"  

# define reusable parameters:
responses:
  UnexpectedError:
    description: "Unexpected error"
    schema:
      $ref: "#/definitions/Error"
  NotFound:
    description: "not found"
    schema:
      $ref: "#/definitions/Error"

definitions:
  Issue:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
      description:
        type: string
      state:
        type: string
        enum: [UNSSIGNED, ASSIGNED, "RESOLVED"]
      userId:
        type: string
      userName:
        type: string        
      agentId:
        type: string
      agentName:
        type: string
      createAt:
        type: string
      assignedAt:
        type: string
      solvedAt:
        type: string
  Agent:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      state:
        type: string
        enum: [free, busy]
  User:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
  Error:
    type: "object"
    properties:
      message:
        type: "string"